refactor: rename tournament-related components and update routing for consistency
