+++ src/pages/GameLobbyPage/TournamentPage/components/MatchNode/styles.css.ts
