+++ src/pages/GameLobbyPage/TournamentPage/components/PlayerCard/styles.css.ts
