+++ src/pages/GameLobbyPage/TournamentPage/index.tsx
