+++ src/pages/GameLobbyPage/TournamentPage/styles.css.ts
